feat(#520): move API token demo into todo-app, drop api-tokens example
Replace the standalone api-tokens example with a direct integration in
the existing todo-app. The JSON API routes (/api/todos GET+POST) are now
protected by RequireApiToken middleware; clients exchange a principal ID
for a Bearer token via POST /api/tokens before calling them. This keeps
example count down while showing a realistic mobile-client scenario.

Fixes the CI failure: examples/api-tokens/Cargo.toml missing workspace
lints (workspace_forbids_unsafe_code_for_all_members) by removing the
example entirely.
