--- examples/api-tokens/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "api-tokens"
-edition.workspace = true
-version.workspace = true
-publish = false
-
-[dependencies]
-autumn-web = { path = "../../autumn" }
-diesel = { version = "2", features = ["postgres", "chrono"] }
-diesel-async = { version = "0.8", features = ["postgres"] }
-pq-sys = { version = "0.7", features = ["bundled_without_openssl"] }
-diesel_migrations = "2"
-serde = { version = "1", features = ["derive"] }
-tokio = { version = "1", features = ["full"] }
-tracing = "0.1"
